MOD:rename trie directory
